Add event listener to header, use *ngIf to conditionally display section
** Inside app.component.html:

* The header has an event listener listening to the custom generated
sectionClicked event, when it's triggered it will call the
onSectionClicked() method that receives the event object as an argument,
it contains the name of the section we want to display.

* Using *ngIf and ng-template we conditionally display either the
Recipes component or the Shopping List component, depending on the
string storend inside the sectionToDisplay property.

* Inside app.component.ts:

* the onSectionClicked() method receives the object and sets the
sectionToDisplay property to whatever string (name of section) was
passed.